[DAT-47318] - Exported renderer functions
