+++ app/src/generate.js
@@ -8,6 +8,7 @@ const excludes = [
   'Select.common.js',
   'select.utils',
   'table.types.js',
+  'Table.common.js',
   'notifications.styles.js',
   'notifications.utils.js',
   'date.utils'
